refactor: move durg class to drug.js
