+++ pharmacy.test.js
@@ -1,9 +1,10 @@
-import { Drug, Pharmacy } from "./pharmacy";
+import { Pharmacy } from "./pharmacy";
+import { Drug } from "./drug";
 
 describe("Pharmacy", () => {
   it("should decrease the benefit and expiresIn", () => {
-    expect(new Pharmacy([new Drug("test", 2, 3)]).updateBenefitValue()).toEqual(
-      [new Drug("test", 1, 2)]
-    );
+    expect(
+      new Pharmacy([new Drug("test", 2, 3)]).updateBenefitValue()
+    ).toEqual([new Drug("test", 1, 2)]);
   });
 });
